chore: add ruff + mypy, wire into CI, and clear what they flagged
Adds ruff (lint + import sort) and mypy to the dev extras and the CI
job so the engine is linted and type-checked, not just tested.

Getting both to green surfaced real dead code, now removed:
- BetaEstimator._unlever_beta: a stub that computed values then always
  returned (None, None, None); the real Hamada unlevering already lives
  in WACCBuilder.compute_wacc. Also dropped a matching dead
  unlevered_beta block there.
- reverse_dcf round-trip check ran a full forward DCF and discarded it.
- sensitivity built a _mock_wacc object never passed downstream (WACC is
  passed straight to _compute_scenario).
- backtest/run computed split_ratio that a later refactor made unused,
  with a now-contradicted comment.
- assorted unused imports/locals across the engine and tests.

Genuine type fixes mypy caught:
- utils.bootstrap_ci annotated statistic_fn as the builtin callable
  instead of typing.Callable.
- PriceFetcher.fetch takes fractional years (0.05) but was typed int.

Config lives in pyproject: __init__ re-export facades and the data.py
compat shim are ignored, mypy runs check_untyped_defs without demanding
full annotation coverage.

+++ openquant/data/providers/prices.py
@@ -27,7 +27,7 @@ class PriceFetcher:
     def fetch(
         self,
         ticker: str,
-        years: int = BETA_LOOKBACK_YEARS,
+        years: float = BETA_LOOKBACK_YEARS,
     ) -> pd.Series:
         """
         Fetch adjusted closing prices for a ticker.
